backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.26.0
apscheduler==3.10.4
aiosqlite==0.19.0
